Cleanup code, remove duplicated
