+++ packages/bodiless-core/src/contextMenuForm.static.tsx
@@ -15,13 +15,6 @@
 import React, { useCallback } from 'react';
 import type { ContextMenuFormProps } from './Types/ContextMenuTypes';
 
-export type Options<D> = {
-  submitValues?: (componentData: D) => boolean|void;
-  onClose?: (componentData: D) => boolean|void;
-  initialValues?: D;
-  hasSubmit?: ((componentData: D) => boolean) | boolean;
-};
-
 export const ContextMenuForm = () => <></>;
 
 export const contextMenuForm = () => () => (
