+++ packages/bodiless-core/src/hoc.bl-edit.tsx
@@ -14,55 +14,17 @@
 import ResizeObserver from 'resize-observer-polyfill';
 
 import React, {
-  ComponentType as CT, EventHandler, FC,
+  ComponentType as CT,
   useEffect,
   useRef,
   useState,
 } from 'react';
-import pick from 'lodash/pick';
 import { HOC } from '@bodiless/fclasses';
 import { observer } from './mobx.bl-edit';
-import { useContextActivator, useExtendHandler, useClickOutside } from './hooks';
+import { useContextActivator } from './hooks';
 import LocalContextMenu from './components/LocalContextMenu.bl-edit';
-
-/**
- * Utility hoc to add an event handler which extends any handler passed to
- * the original component.
- *
- * Only adds the extension when in edit mode.
- *
- * @param event The name of the event whose handler is to be extended
- * @param useExtender Custom hook returning the handler to add. Will be invoked
- *        during render and receive the original props of the component.
- *
- * @return An HOC which will add the handler.
- */
-export const withExtendHandler = <P extends object>(
-  event: string,
-  useExtender: (props: P) => EventHandler<any>,
-): HOC => Component => {
-    const WithExtendHandler: FC<any> = props => (
-      <Component
-        {...props}
-        {...useExtendHandler(event, useExtender(props), props)}
-      />
-    );
-    return WithExtendHandler;
-  };
-
-/*
- * Creates an HOC which strips all but the specified props.
- *
- * @param keys A list of the prop-names to keep.
- *
- * @return An HOC which will strip all but the specified props.
- */
-export const withOnlyProps = <Q extends object>(...keys: string[]) => (
-  <P extends object>(Component: CT<P> | string) => {
-    const WithOnlyProps: FC<P & Q> = props => <Component {...pick(props, keys) as P} />;
-    return WithOnlyProps;
-  }
-);
+import { withClickOutside, withExtendHandler, withOnlyProps } from './hoc.static';
+import type { resizeDetectorProps, ClickOutsideProps } from './hoc.static';
 
 /**
  * Creates an HOC which provides the base component event handler which activates the current
@@ -100,43 +62,6 @@ export const withLocalContextMenu: HOC = Component => {
     </LocalContextMenu>
   );
   return WithLocalContextMenu;
-};
-
-export type ClickOutsideProps = {
-  onClickOutside?: (e: KeyboardEvent | MouseEvent) => void;
-};
-
-/**
- * Utility hoc to add onClickOutside handler to the original component.
- * A callback will be executed on both click outside as well as on the `esc` keypress.
- *
- * @return An HOC which will add the handler.
- */
-export const withClickOutside = <P extends object>(Component: CT<P> | string) => {
-  const WithClickOutside = (props: P & ClickOutsideProps) => {
-    const { onClickOutside } = props;
-    const ref = useRef(null);
-
-    // Only add listners if onClickOutside handler is defined
-    if (typeof onClickOutside === 'function') {
-      useClickOutside(ref, onClickOutside);
-    }
-
-    return (
-      <div ref={ref}>
-        <Component {...props} />
-      </div>
-    );
-  };
-
-  return WithClickOutside;
-};
-
-export type resizeDetectorProps = {
-  onResizeObserver?: (
-    ref:React.MutableRefObject<any>,
-    entries: ResizeObserverEntry[],
-  ) => void;
 };
 
 /**
@@ -179,4 +104,15 @@ export const withResizeDetector = <P extends object>(Component: CT<P> | string) 
   };
 
   return WithResizeDetector;
+};
+
+export {
+  withOnlyProps,
+  withExtendHandler,
+  withClickOutside,
+};
+
+export type {
+  ClickOutsideProps,
+  resizeDetectorProps,
 };
